Implement getting the users student location info.
-Dispatch getting the users student location after login (aka post a
session function)
-In getPublicUserData function, I took the JSON dictionary of newData
because for security purposes Udacity wants students to use a subset of
data without the five first characters.
-In getAStudentLocation function, I applied percent encoding with
urlQueryAllowed for stringURL to become NSURL. Without it the app will
crash because some characters are not valid URL characters.
-Also in getAStudentLocation function, I take the first object of the
resulting JSON array to add to shared model as the logged in User’s
student location.